Added new ways to create, edit, and delete accounts.
