+++ lib/constants/bottom_tab_navigators.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:growth/constants/nav_routes.dart';
 
-import 'package:growth/pages/account/account_info_page.dart';
+import 'package:growth/pages/account/account_edit_page.dart';
+import 'package:growth/pages/account/account_create_page.dart';
 
 import 'package:growth/pages/balance/balance_details_page.dart';
 import 'package:growth/pages/balance/balance_home_page.dart';
@@ -28,9 +29,11 @@ class BottomTabNavigators {
             case NavigationRoutes.balanceDetailsRoute:
               builder = (BuildContext context) => const BalanceDetailsPage();
               break;
-            case NavigationRoutes.balanceAccount:
-              builder =
-                  (BuildContext context) => const AccountInfoPage();
+            case NavigationRoutes.balanceAccountCreate:
+              builder = (BuildContext context) => const AccountCreatePage();
+              break;
+            case NavigationRoutes.balanceAccountEdit:
+              builder = (BuildContext context) => const AccountEditPage();
               break;
 
             default:
